# openapi/gen/bundled/auth.yaml
openapi: 3.0.0
info:
  title: Authentication API
  version: '1.0'
  contact:
    name: support@xata.io
...
paths:
  /organizations:
    summary: Organizations Management
    description: This endpoint enables creating and listing organizations within Xata. Organizations are the top-level entity that contains projects, branches, and team members.
    get:
      operationId: getOrganizationsList
      summary: Get list of organizations
      description: Retrieve the list of all organizations the authenticated user belongs to, including their IDs and names. This endpoint allows users to view all organizations they have access to.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  organizations:
                    description: List of organizations the user has access to
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          description: Unique identifier for the organization
                          $ref: '#/components/schemas/OrganizationID'
                        name:
                          description: Human-readable name of the organization
                          type: string
                        status:
                          description: Current status of the organization
                          $ref: '#/components/schemas/OrganizationStatus'
                      required:
                        - id
                        - name
                        - status
                required:
                  - organizations
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/AuthError'
        '404':
          $ref: '#/components/responses/SimpleError'
        5XX:
          description: Unexpected Error
      security:
        - xata:
            - org:read
      tags:
        - Organizations
    post:
      operationId: createOrganization
      summary: Create a new organization
      description: Creates a new organization with the authenticated user as its single owner.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationRequest'
      responses:
        '201':
          description: Organization successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/AuthError'
        '404':
          $ref: '#/components/responses/SimpleError'
        5XX:
          description: Unexpected Error
      security:
        - xata:
            - org:write
      tags:
        - Organizations
  /organizations/{organizationID}:
    summary: Organization Operations
    description: This endpoint provides operations for a specific organization, allowing retrieval, updating, and deletion of an organization by its unique identifier.
    parameters:
      - $ref: '#/components/parameters/OrganizationIDParam'
    get:
      operationId: getOrganization
      summary: Get organization details
      description: Retrieve detailed information about a specific organization by its ID.
      responses:
        '200':
          description: Organization details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          $ref: '#/components/responses/AuthError'
        '404':
          $ref: '#/components/responses/SimpleError'
        5XX:
          description: Unexpected Error
      security:
        - xata:
            - org:read
      tags:
        - Organizations
    put:
      operationId: updateOrganization
      summary: Update organization details
      description: Update information for an existing organization, such as its name.
      requestBody:
        content:
          application/json:
            schema:
              properties:
                id:
                  description: Unique identifier for the organization to update
                  $ref: '#/components/schemas/OrganizationID'
                name:
                  description: New name for the organization
                  type: string
      responses:
        '200':
          description: Organization successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          $ref: '#/components/responses/AuthError'
        '404':
          $ref: '#/components/responses/SimpleError'
        5XX:
          description: Unexpected Error
      security:
        - xata:
            - org:write
      tags:
        - Organizations
    delete:
      operationId: deleteOrganization
      summary: Delete an organization
      description: '[Not implemented yet] Permanently delete an organization. This action cannot be undone.'
      responses:
        '204':
          description: Organization successfully deleted
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          $ref: '#/components/responses/AuthError'
        '404':
          $ref: '#/components/responses/SimpleError'
        5XX:
          description: Unexpected Error
      security:
        - xata:
            - org:write
      tags:
        - Organizations
  /organizations/{organizationID}/api-keys:
    get:
      operationId: listOrganizationAPIKeys
      summary: List API Keys for an organization
      description: Retrieve a list of API keys for a specific organization.
      parameters:
        - $ref: '#/components/parameters/OrganizationIDParam'
      responses:
        '200':
          description: List of API keys retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys:
                    description: Array of API keys for the organization
                    type: array
                    items:
                      $ref: '#/components/schemas/APIKeyPreview'
                required:
                  - keys
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          $ref: '#/components/responses/AuthError'
        '404':
          $ref: '#/components/responses/SimpleError'
        5XX:
          description: Unexpected Error
      security:
        - xata:
            - keys:read
      tags:
        - API Keys
    post:
      operationId: createOrganizationAPIKey
      summary: Create an Organization API Key
      description: Create a new API key for an organization.
      parameters:
        - $ref: '#/components/parameters/OrganizationIDParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAPIKeyRequest'
      responses:
        '201':
          description: API Key created
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    $ref: '#/components/schemas/FullAPIKey'
                required:
                  - key
        '400':
          description: API key limit reached or invalid request
      security:
        - xata:
            - keys:write
      tags:
        - API Keys
    delete:
      operationId: deleteOrganizationAPIKeys
      summary: Bulk delete API Keys for an organization
      description: Delete multiple API keys in a single request.
      parameters:
        - $ref: '#/components/parameters/OrganizationIDParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  description: Array of API key IDs to delete
                  type: array
                  items:
                    type: string
                  maxItems: 50
              required:
                - ids
      responses:
        '204':
          description: API Keys deleted successfully
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/AuthError'
        '404':
          $ref: '#/components/responses/SimpleError'
      security:
        - xata:
            - keys:write
      tags:
        - API Keys
  /organizations/{organizationID}/members:
    get:
      operationId: listOrganizationMembers
      summary: List members of an organization
      parameters:
        - $ref: '#/components/parameters/OrganizationIDParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserWithID'
                required:
                  - members
      security:
        - xata:
            - org:read
      tags:
        - Organizations
  /organizations/{organizationID}/members/{userID}:
    delete:
      operationId: removeOrganizationMember
      summary: Remove a member from an organization
      parameters:
        - $ref: '#/components/parameters/OrganizationIDParam'
        - $ref: '#/components/parameters/UserIDParam'
      responses:
        '204':
          description: Member removed
      security:
        - xata:
            - org:write
      tags:
        - Organizations
  /organizations/{organizationID}/invitations:
    get:
      operationId: listOrganizationInvitations
      summary: List invitations for an organization
      description: Retrieve all invitations for an organization with optional filtering by status, email, and name.
      parameters:
        - $ref: '#/components/parameters/OrganizationIDParam'
        - name: status
          in: query
          description: Filter invitations by status
          schema:
            type: string
            enum:
              - pending
              - expired
        - name: email
          in: query
          description: Filter invitations by email address
          schema:
            type: string
        - name: first_name
          in: query
          description: Filter invitations by first name
          schema:
            type: string
        - name: last_name
          in: query
          description: Filter invitations by last name
          schema:
            type: string
        - name: search
          in: query
          description: Search invitations by email or name
          schema:
            type: string
        - name: first
          in: query
          description: Index of the first result to return (0-based offset for pagination)
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: max
          in: query
          description: Maximum number of results to return
          schema:
            type: integer
            default: 100
            maximum: 100
            minimum: 1
      responses:
        '200':
          description: List of invitations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  invitations:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrganizationInvitation'
                required:
                  - invitations
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          $ref: '#/components/responses/AuthError'
        '404':
          $ref: '#/components/responses/SimpleError'
        5XX:
          description: Unexpected Error
      security:
        - xata:
            - invite:read
      tags:
        - Organizations
    post:
      operationId: createOrganizationInvitation
      summary: Send an invitation to join an organization
      description: Send an invitation email to a user to join an organization.
      parameters:
        - $ref: '#/components/parameters/OrganizationIDParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationInvitationRequest'
      responses:
        '201':
          description: Invitation sent successfully
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          $ref: '#/components/responses/AuthError'
        '404':
          $ref: '#/components/responses/SimpleError'
        '409':
          $ref: '#/components/responses/SimpleError'
        5XX:
          description: Unexpected Error
      security:
        - xata:
            - invite:write
      tags:
        - Organizations
...
  /organizations/{organizationID}/membership-limits:
    get:
      operationId: getOrganizationMembershipLimits
      summary: Get organization membership limits
      description: Retrieves the membership limits for the specified organization (maximum members and invitations).
      parameters:
        - $ref: '#/components/parameters/OrganizationIDParam'
      responses:
        '200':
          description: Membership limits for the organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationMembershipLimits'
        '401':
          $ref: '#/components/responses/AuthError'
        '403':
          $ref: '#/components/responses/AuthError'
        5XX:
          description: Unexpected Error
      security:
        - xata:
            - org:read
      tags:
        - Organizations
...
components:
...
  schemas:
...
    CreateOrganizationRequest:
      description: Request payload for creating a new organization
      type: object
      properties:
        name:
          description: Name for the new organization
          type: string
      required:
        - name
...
    OrganizationID:
      title: OrganizationID
      type: string
      pattern: '[a-zA-Z0-9_-~:]+'
      x-oapi-codegen-extra-tags:
        validate: identifier
  responses:
    BadRequestError:
      description: Error returned when the request is malformed or contains invalid parameters
      content:
        application/json:
          schema:
            type: object
            properties:
              id:
                description: Error identifier for tracking and debugging
                type: string
              message:
                description: Human-readable error message explaining the issue
                type: string
            required:
              - message
    AuthError:
      description: Error returned when authentication or authorization fails
      content:
        application/json:
          schema:
            type: object
            properties:
              id:
                description: Error identifier for tracking and debugging
                type: string
              message:
                description: Human-readable error message explaining the issue
                type: string
            example:
              message: invalid API key
            required:
              - message
...